modifier: Added modifier script
Currently capable of extracting all a color PD3 file into a folder and
to combine it afterwards back into the original file.

Extract stores all relevant image files in the output folder together
with a json file, which describes the header data.

Combine retrieves these instances, does a simple verification on it and
afterwards stores the file again.

So let's go modify the heck out of it ^^

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>