color: modify: Add verify command to script
There's really no reason to have the modify script with all it's
capabilities and have the verify script in its own file. This was
initially done as verification was build before the manipulation
tool-set, but as all capabilities from the verifier has now been
repacked into the lib file, we can simply integrate it into one
universal script for COLOR PD3s.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>